refactor: modify fields for mapping correctly
