--- BE/src/main/java/com/codesquad/issue04/domain/entity/User.java
@@ -1,26 +1,0 @@
-package com.codesquad.issue04.domain.entity;
-
-
-import lombok.*;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Getter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Builder
-@Entity
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    private String githubId;
-    private String image;
-
-}
