+++ BE/src/main/java/com/codesquad/issue04/domain/milestone/Milestone.java
@@ -1,10 +1,15 @@
-package com.codesquad.issue04.domain.entity;
+package com.codesquad.issue04.domain.milestone;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import javax.persistence.OneToMany;
 
+import com.codesquad.issue04.domain.issue.Issue;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -13,12 +18,15 @@ import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 @Getter
-@ToString
+@ToString(exclude = "issues")
 @NoArgsConstructor
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Entity
 @Builder
 public class Milestone {
 
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	private String title;
 	private LocalDateTime dueDate;
